created basepage class and fixed inheritance structure
